requirements.txt
uvicorn
requests
pandas
scikit-learn
xgboost
nest_asyncio
joblib
fastapi
python-dotenv
pydantic
